+++ Source/Backends/Vulkan/TestDevice/Source/Device.cpp
@@ -816,7 +816,7 @@ ResourceSetID Device::CreateResourceSet(ResourceLayoutID layout, const ResourceI
     return Test::ResourceSetID(static_cast<uint32_t>(resourceSets.size()) - 1);
 }
 
-PipelineID Device::CreateComputePipeline(const ResourceLayoutID* layouts, uint32_t layoutCount, const void *shaderCode, uint32_t shaderSize) {
+PipelineID Device::CreateComputePipeline(const ResourceLayoutID* layouts, uint32_t layoutCount, const void *shaderCode, uint64_t shaderSize) {
     PipelineInfo& info = pipelines.emplace_back();
 
     // Shader info
